requirements.txt
tensorflow==2.15.0
keras==2.15.0          # keep tf/keras in sync
streamlit==1.34.0
numpy
pillow
gdown
